MCR-1349 #comment change type of getMaxUploadSize() from int to long

+++ src/main/java/org/swordapp/server/SwordConfiguration.java
@@ -19,7 +19,7 @@ public interface SwordConfiguration {
 
     String getTempDirectory();
 
-    int getMaxUploadSize();
+    long getMaxUploadSize();
 
     String getAlternateUrl();
